made all the modules fall under lib.rs [great for testing later on]
basically we use a lib.rs file to refer to our modules so that the main
program can refer to the modules using the "aleph"::<module_name>
modifier. This lets us do the same in the tests, quite nifty.

additionally I have made all the members of the manifest crate public